move remembver function out of message composer:
